refactor: course-detail の God コンポーネントを ui/lib へ分解し命名を可読化 (FRESTYLE-208)
CourseDetailPage.tsx(879行) に同居していた 8 コンポーネントと 3 ユーティリティを
FSD の page-local な ui/ lib/ へ分解（1 ファイル 1 コンポーネント）。振る舞いは不変。

- ui/: MaterialListItem / ManagedDetail / ReadOnlyDetail / ChapterNav /
  MaterialSkeleton / CompleteToggleButton / ImageLightbox / ReadOnlyMarkdown
- lib/: formatDate(+padTwoDigits) / stripLeadingTitle
- ファイル途中の import 文を各コンポーネント先頭へ集約
- 省略名を意味の分かる名前へ改名: s→state / idx→selectedIndex / m→material /
  mid→materialId / ok→succeeded / e→event / i→index / v→isOpen / c→fetchedCourse /
  active→isCurrent / n→value / d→date / closeRef→closeButtonRef / onKey→handleKeyDown

CourseDetailPage は 879→309 行。既存テストは assert 不変で全通過（DOM 出力不変を担保）。
tsc / vitest(1221) / eslint / build / branches 80.17%(閾値維持) すべて green。
